+++ src/v2/Components/ArtworkFilter/SavedSearch/types.ts
@@ -26,3 +26,18 @@ export interface SavedSearchAttributes {
 }
 
 export type SearchCriteriaAttributeKeys = keyof SearchCriteriaAttributes
+
+export interface DefaultFilterPill {
+  isDefault: true
+  name: string
+  displayName: string
+}
+
+export interface NonDefaultFilterPill {
+  isDefault?: false
+  name: string
+  displayName: string
+  filterName: string
+}
+
+export type FilterPill = DefaultFilterPill | NonDefaultFilterPill
